Update fair-principles.md
Initial feedback on the FAIR principles recipe. Thx to @dcfcjcheadle for the fast turn around. @u8sand @AviMaayan I have added the links and references to FAIRShake and reworded some of the section. We can still tweak the word-smithing.

TODO: mesh this with the specific recipes about the tools themselves and the evaluation of the DCC datasets.